fix(api): rename typeorm folder
